1차 완료. 이제 아이콘 하고 readme 를 좀 이쁘게 포장해서 배포 준비

+++ lib/component/Option_Tts.dart
@@ -210,7 +210,8 @@ class Option_Tts extends OptionsBase {
                                               1,
                                           onChanged: (value) {
                                             (controller.config['tts']
-                                                as Map)['groupcnt'] = value;
+                                                    as Map)['groupcnt'] =
+                                                value.toInt();
                                             controller.update();
                                           },
                                         ),
@@ -240,7 +241,6 @@ class Option_Tts extends OptionsBase {
   @override
   Widget build(BuildContext context) {
     // TESTopenSetting();
-    // print('tts load--------');
     return IconButton(
         onPressed: () {
           openSetting();
